requirements.txt
# --- For-Bot-Working --------- #
pyrogram==1.1.2
Telethon~=1.23.0
pyromod==1.5
Pillow
python-dotenv
pyshorteners
TgCrypto
heroku3
wget
motor
python-arq
dnspython
bs4
gitpython
psutil
python-arq
gTTS
asyncio
googletrans==4.0.0-rc1
youtube-dl
youtube_search
requests
aiohttp
# --- For-Database ------------ #
sqlalchemy~=1.3.23
psycopg2-binary
feedparser
tswift
img2pdf


